--- src/day11/Dortgen.java
@@ -1,21 +1,0 @@
-package day11;
-
-public class Dortgen {
-    public static void main(String[] args) {
-        //TODO: Dortgen programını geliştiriniz.
-        /*
-            1. Klavyeden 3 adet girdi alınacaktır.
-                kenar:int       -> Karenin kenarı
-                kisaKenar:int   -> Dikdörtgenin kısa kenarı
-                uzunKenar:int   -> Dikdörtgenin uzun kenarı
-            2. alanHesapla(int k1, int k2):int  -> Metodu geliştirilecektir.
-            3. Tek bir metot (alanHesapla) iki kez çağırılarak hem karenin hem de
-            dikdörtgenin alanı hesaplanacaktır.
-            4. cevreHesapla(int k1,int k2):int  -> Metodu geliştirilecektir.
-            5. Tek bir metot (cevreHesapla) iki kez çağırılarak hem karenin hem de
-            dikdörtgenin çevresi hesaplanacaktır.
-            6. Elde edilen toplam alan hesaplanıp ekranda gösterilecektir.
-            7. Elde edilen toplam çevre ekranda gösterilecektir.
-         */
-    }
-}
